requirements.txt
pixivpy-async>=1.2.14
aiohttp>=3.9.0
aiosqlite>=0.19.0
pyyaml>=6.0
python-telegram-bot>=20.7
apscheduler>=3.10.4
aiocqhttp>=1.4.4
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
requests>=2.31.0
openai>=1.0.0
Pillow>=10.0.0

selenium>=4.10.0
webdriver-manager>=4.0.0
